move png draw function into module
